Revert "REFACTOR: Make a DeleteMenuItem"
This reverts commit c23657c5343315b912c393c41f4123664bc0786c.
